+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Added
 
+- #2238 Add a "Reduce app killing" card to Expert Mode on Xiaomi/Redmi/Poco devices with steps to whitelist the app from battery optimisation, disable MIUI optimization, enable autostart, and adjust battery saver settings.
 - Target Android 17 SDK.
 - [#2227](https://github.com/keymapperorg/KeyMapper/issues/2227) Add a step to the Expert Mode setup wizard to grant local network access permission, required for ADB on Android 17+.
 - #2223 Add one-finger and two-finger double-tap gestures to the TalkBack gesture action.